+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `Validator` static class is replaced by the namespace `Form\Rule`
 - Using PSR-4 autoloader
 - Adding syntax to access nested fields directly in `getRules`, `getErrors` and `getValues` (and similar `has*` methods)
+- `setRules` now accepts another validator
 
 ### Rule changes